Work in Progress: Exporting C4D Materials and Shaders
Exporting C4D Materials and Shaders to NSI Shaders. First the material with all
its parameters is exported as the main shader for NSI, and then for each shader
used on that material we export them again with its attributes as an NSI shader
and then connected them to the corresponding one.

The attributes’ names for material and shaders in Cinema4D are located in .str
file which makes it not possible to get this names using the code unless they
are in the default c4d_string.str file which would cause an unstructured
behavior if we put all the attributes in there.
Thus we came to a solution to use C4D attributes’ Ids to represent the
attributes’ names in our OSL shaders. With a slight modification the names for
the attributes in our shaders will be underscore plus the ID (“_”+std::string(Id))

In order to find what these attributes are used for together with the ID that
they have, we just go to the corresponding .h file example(Mmaterial.h for
materials) or (Xcheckerboard.h for the checkerboard shader and so)

MATERIAL_COLOR_SHADER = 8000,  (_8000: attribute name in OSL shader)
MATERIAL_DIFFUSION_SHADER = 8001, (_8001: attribute name in OSL shader)
MATERIAL_LUMINANCE_SHADER = 8002, (_8002: attribute name in OSL shader)
MATERIAL_TRANSPARENCY_SHADER = 8003, (_8003: attribute name in OSL shader)
MATERIAL_REFLECTION_SHADER 	= 8004, (_8004: attribute name in OSL shader)

Above are some attributes Ids from the Mmaterial’s header file of C4D.Here we
can clearly see for what these attributes are used, or, if we want to check the
names based on the UI we will have to look at the corresponding .str file where
we will find the ID and name used on the UI for that ID.

TO DO:
•Create OSL shader for the main shader (c4d material) and all the other shaders
•Remove the unused attributes of the shaders (if not ignore them)

FIXME:
•In the connection between the shaders and the main shader (material) the source
 parameter is denoted as “output”. Instead there should be the parameters passed
 from the source shader.