Changed the method of writing 3Delight materials (Using only general translator)
Our old way of dealing with 3Delight materials was by writing a new translator
for each material that we would create. As it was time consuming, we are using
NSIExportMaterial as our general translator for all the new created materials.
We are using a map on NSIExportMaterial which has the Ids and a pair of
datatype and name of each attribute of the material. Since we are getting the
attributes’ names as IDs like we have mentioned in previous commits we should
use the same method for the newly created materials that’s why we are using
this map on NSIExportMaterial and then initializing it on the newly created
shader class. An example of how it will look like would be as below.

On NSIExportMaterial
    std::map<int, std::pair<std::string,std::string> > m_ids_to_names;

On the new created material (This case DLGlass)
    m_ids_to_names[REFLECTION_COLOR] = pair("Cout", "reflect_color");
    m_ids_to_names[REFLECTION_COLOR_SHADER] = pair("Cout", "reflect_color");
    m_ids_to_names[REFLECTION_ROUGHNESS] = pair("Cout[0]", "reflect_roughness");
This way all the created materials are successfully translated on NSI.