Exported Sky Object from Cinema4D and did some minor changes
Exported Sky Object from Cinema4D as an environment shader in 3Delight.
You can connect materials to the sky object and use hdri texture on
that material to get an hdri light in Cinema4D and 3Delight.
Also created a Sky shader for 3Delight which can be connected with the
environment light.
Removed the 3Delight lights and shaders from Cinema4D as we will be
supporting their own light and shader.

To Do:
 • Get a shadow view according to hdri light when it is used. Issue #23
 • Create a material to handle the Sky Object from Cinema4D. Issue #24
 • Remove Dl_Environment shader having a complete support of
   cinema4D hdri light. Issue #25